requirements.txt
asgiref==3.5.0
django-environ==0.8.1
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
sqlparse==0.4.2